+++ packages/media-editor/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Enhancements
+
+-   Show the selected crop aspect ratio in a native select in the compact image controls ([#82969](https://github.com/WordPress/gutenberg/pull/82969)).
+
 ## 0.19.0 (2026-09-23)
 
 ### Bug Fixes